# IP.Prefix: speed up by removing type switch and using fewer ops for masking.