--- a/netaddr_test.go
+++ b/netaddr_test.go
@@ -532,6 +532,7 @@ func TestIPPrefixMasking(t *testing.T) {
 			subtests: []subtest{
 				{
 					bits: 255,
+					ok: true,
 				},
 				{
 					bits: 16,