@danderson 2020-06-07
Removed helper and specialized each impl separately. IPv4 is now a 3-instruction loop (yay elided bounds check), IPv6 invokes memclr.

No idea why /128+zone regresses with this. Do we end up invoking memclr unconditionally on a 0-byte slice?...

```
    name                              old time/op    new time/op    delta
    IPPrefixMasking/IPv4_/32-8          29.4ns ± 2%    27.6ns ± 3%   -6.06%  (p=0.008 n=5+5)
    IPPrefixMasking/IPv4_/17-8          32.8ns ± 2%    30.1ns ± 2%   -8.06%  (p=0.008 n=5+5)
    IPPrefixMasking/IPv4_/0-8           32.1ns ± 2%    31.0ns ± 9%     ~     (p=0.143 n=5+5)
    IPPrefixMasking/IPv6_/128-8         43.2ns ± 3%    40.8ns ± 1%   -5.55%  (p=0.008 n=5+5)
    IPPrefixMasking/IPv6_/65-8          47.3ns ± 1%    45.9ns ± 2%   -2.83%  (p=0.032 n=5+5)
    IPPrefixMasking/IPv6_/0-8           52.2ns ± 1%    49.1ns ± 3%   -5.83%  (p=0.008 n=5+5)
    IPPrefixMasking/IPv6_zone_/128-8    76.5ns ± 3%    86.4ns ± 5%  +12.88%  (p=0.008 n=5+5)
    IPPrefixMasking/IPv6_zone_/65-8     92.8ns ± 3%    86.5ns ± 2%   -6.73%  (p=0.008 n=5+5)
    IPPrefixMasking/IPv6_zone_/0-8      98.5ns ± 3%   101.1ns ± 6%     ~     (p=0.206 n=5+5)
```