@danderson 2020-06-07
Going to merge like this. I'm not 100% convinced about the IPv4 impl that avoids memclr, because that'll become less good if the compiler gets better and inlines memclr intelligently. But as of Go 1.14, this is faster, so, shrug?